Moving HeapSizeManager and TestHeapSizeManager to bigtable-client-core. I used eclipse's refactoring to extract the HeapSizeManager class.  I had to make a couple of methods public, but that was the only changes that I made to the code itself.
